ProductDesigner/requirements.txt
langchain>=0.1.0
langgraph>=0.0.20
langchain-openai>=0.0.2
langchain-community>=0.0.10
pydantic>=2.5.2
typing-extensions>=4.8.0
python-dotenv>=1.0.0
duckduckgo-search>=3.9.5
openai>=1.3.0
flask>=2.3.0
flask-socketio>=5.3.0
python-socketio>=5.8.0
